// extensions/xcode/src/models/xcode-simulator/xcode-simulator-app-privacy-service-type.model.ts
/**
 * A Xcode Simulator App Privacy Service Type
 */
export enum XcodeSimulatorAppPrivacyServiceType {
  all = "all",
  calendar = "calendar",
  contactsLimited = "contacts-limited",
  contacts = "contacts",
  location = "location",
  locationAlways = "location-always",
  photos = "photos",
  photosAdd = "photos-add",
  mediaLibrary = "media-library",
  microphone = "microphone",
  motion = "motion",
  reminders = "reminders",
  siri = "siri",
}